bookhub-main/requirements_personal.txt
asgiref
Django
django-crispy-forms
psycopg2
pytz
sqlparse
typing-extensions
crispy-bootstrap5
